Fixed some bugs related to closing the file using "X"
When we enter an invalid number , we call the function again
Now when we try to exit the function by typing "x" ,
sys.exit() exits only from the current function.
It doesn't stop the program instead. So instead of using "x" to exit ,
we just inform the user to use "Ctrl+x" to exit the program.