Added save function to my_text_editor.py. The program has been completed!
The program can save files in .docx or .txt or .csv extension only.
All files option is not made available during saving because it may
corrupt any file such as VLC or even PDF as file.write() is writing
its contents to those files also!

So All Files option is not made available for the users!